feat: 实现短链接生成完整流程，添加过滤词表和修复hash溢出问题
- 添加短链接过滤词表配置（技术保留词、敏感词、品牌词等）
- 实现短链生成逻辑，支持特殊词汇过滤
- 修复 lurl_hash 字段类型溢出问题（BIGINT UNSIGNED -> SIGNED）
- 添加数据库迁移文件
- 完善 SequenceDB 配置项
- 添加 Domain 配置用于拼接短链接

+++ model/shorturlmapmodel.go
@@ -38,7 +38,7 @@ func (m *customShortUrlMapModel) withSession(session sqlx.Session) ShortUrlMapMo
 
 func (m *customShortUrlMapModel) FindByHashAndLurl(ctx context.Context, lurl string, hashvalue uint64) (*ShortUrlMap, error) {
 	// MurmurHash3
-	lurlHash := hashvalue
+	lurlHash := int64(hashvalue)
 
 	// 查询数据库
 	sqlStr := fmt.Sprintf("select %s from %s where `lurl_hash` = ? and `lurl` = ? limit 1", shortUrlMapRows, m.table)
